Characters can no longer break the program by trying to go into the void. Next up, make sure windows dont overlap with doors. Then make doors on adjacent rooms line up and lead to each other.

+++ assignments/4/4.0-room-renderer/RoomRenderer/RoomRenderer/Program.cs
@@ -31,7 +31,7 @@ namespace RoomRenderer
             //var renderer = new Renderer();
             renderer.RenderContainer(container);
             var manager = new CharacterManager(1, 1, container.AllRenderPoints);
-            while (manager.CharacterLooper((int)Console.ReadKey(true).Key)) { }
+            while (manager.CharacterLooper(Console.ReadKey(true).Key)) { }
         }
 
         public void TestingBoundaries()
